+++ netharn/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
-__version__ = '0.0.26'
+__version__ = '0.0.27'
 """
 mkinit netharn
 """